pyproject.toml
[project]
name = "tgbot-python-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.3",
    "openai>=1.58.1",
    "pygithub>=2.5.0",
    "python-telegram-bot[job-queue]>=21.9",
    "pytz>=2024.2",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "isort>=5.13.2",
]

[tool.isort]
profile = "black"

[tool.black]
line-length = 120
